libmpathpersist: Don't try release workaround for invalid type
When trying to release a reservation, if the user specified the wrong
reservation type, libmpathpersist would try to preempt the reservation,
because the reservation key matched the device key, but it was not
removed.  In this case, the preemption would also fail because it also
requires a matching type.

Check if the reservation type matches, to avoid attempting the
workaround in this case.